pyproject.toml
[tool.poetry]
name = "serverino"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.3"
fastapi = "^0.68.1"
geopandas = "^0.9.0"
Shapely = "^1.7.1"
Fiona = "^1.8.20"
uvicorn = "^0.15.0"
Rtree = "^0.9.7"
geojson-pydantic = "^0.3.1"
sentinelhub = "^3.4.0"
rasterio = "^1.2.8"
aiofiles = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
